authentication and authorization

+++ app/Http/Controllers/IdeaController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Idea;
+use Auth;
 use Illuminate\Http\Request;
 
 
@@ -20,12 +21,20 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea)
     {
+        if(auth()->id()!== $idea->user_id)
+        {
+            abort('404');
+        }
         $editing = true;
         return view('ideas.show', compact('idea', 'editing'));
     }
 
     public function update(Idea $idea)
     {
+        if(auth()->id()!== $idea->user_id)
+        {
+            abort('404');
+        }
         $validated = Request()->validate(
             [
                 'content' => 'Required|min:3|max:240'
@@ -43,12 +52,16 @@ class IdeaController extends Controller
                 'content' => 'Required|min:3|max:240'
             ]
         );
+        $validated['user_id']=auth()->id();
         Idea::create($validated);
         return redirect()->route('dashboard')->with('success', 'idea created successfully');
     }
     public function destroy(Idea $idea)
     {
-
+        if(auth()->id()!== $idea->user_id)
+        {
+            abort('404');
+        }
         $idea->delete();
         return redirect()->route('dashboard')->with('success', 'idea deleted successfully');
     }
